+++ lizard_map/media/lizard_map/lizard_map.js
@@ -247,7 +247,7 @@ function setUpWorkspaceAcceptable() {
                             extent.west, extent.north,
                             extent.east, extent.south);
                         map.setCenter(ol_extent.getCenterLonLat(),
-                                      map.getZoomForExtent(ol_extent));
+                                      map.getZoomForExtent(ol_extent) - 3);
                     }
                 });
             });
@@ -475,7 +475,7 @@ function setUpWorkspaceItemPanToLayer() {
                         extent.west, extent.north,
                         extent.east, extent.south);
                     map.setCenter(ol_extent.getCenterLonLat(),
-                                  map.getZoomForExtent(ol_extent) - 1);
+                                  map.getZoomForExtent(ol_extent) - 3);
                 }
             });
     });